BUG: Fix incorrect cython definition of npy_cfloat
The definition in the numpy headers is a pair of `float`s not `double`s.

+++ numpy/__init__.cython-30.pxd
@@ -329,8 +329,8 @@ cdef extern from "numpy/arrayobject.h":
     ctypedef long double  npy_float128
 
     ctypedef struct npy_cfloat:
-        double real
-        double imag
+        float real
+        float imag
 
     ctypedef struct npy_cdouble:
         double real
